Added support for documentation and operations plugins
