--- bonfire.conf
@@ -1,5 +1,0 @@
-
-
-# Verbose level
-[Verbose]
-Level: 1
