+++ lib/plugins/alerts/AWS.py
